Refactor for ergonomics and correctness
It is now clear to me that the `secs::Secs` should be created within the
context of a `sig::Signature` (that is: SIGSTRUCT). The workflow is that
the enclave loader will load the signature (SIGSTRUCT) for a binary and
then use the data contained therein to build an SECS page.