Add tests to the hasher
First we made the hasher produce the same output as the Linux kernel SGX
selftest. Then we made a bunch of tests to verify the algorithm doesn't
accidentally change. These tests have not yet been tested against the Intel
CPU, so they might still be incorrect in some corner cases. However, the
algorithm should be fairly correct at this point.