Don't expose fields on sgx-types::attr::Attributes
In an ideal world, the Attributes struct would not need to be packed. Then
we could just make the fields public. The problem arises in Signature
(SIGSTRUCT). There we find an unforunate layout. The data inside Signature
that needs to be signed is non-contiguous. We group the two regions
together under the Author and Content structs, respectively. Unfortunately,
the first field of the Contents struct is aligned to 4 bytes in Signature.
This means that the alignment of Contents must be no larger than 4 bytes.
However, the alignment of Attributes is 8 bytes. To get around this, we
ensure that Attributes is packed.

Unfortunately, we can't take the address of a packed struct field. Which
means that none of the bitfield methods will work if called directly on the
fields themselves. Having getters for these fields forces the compiler to
generate a separate load, the results of which are properly aligned. This
is both more ergonomic and doesn't require unsafe code.