Guide readers to "How to contribute" on the wiki
This commit is accompanied by a corresponding change to the Enarx
wiki page titled "How to contribute"
(https://github.com/enarx/enarx/wiki/How-to-contribute). Since setting
up a development environment is covered by that wiki page change, let's
remove the duplicated information recorded in the README so we only
maintain it in one place (the wiki) and have the README point readers
to that wiki page.