Mark secs::Spec::max_enc_size() as safe
This function is technically unsafe because it does not check if the
instruction is available before issuing it. This could result in a crash
on some very old CPUs. However, the only modern context where it could
crash is environments like SGX or some virtualized CPUs. But even in
these contexts, it is common to trap and emulate the instruction.

Therefore, it is common practice to ignore the test to see if the
instruction is available and just issue it anyway. For this reason, we
are marking this function as safe. For more background to the state of
CPUID in Rust, see:

https://github.com/rust-lang/rust/issues/60123