Improve the per-crate license test
First, use the top-level LICENCE file as the canonical license. This
avoids network access during tests.

Second, use diff instead of shell equality to test for differences in
the texts. This allows us to ignore whitespace and newline changes. It
also has the advantage of showing the actual differences during a
failure.

Third, shorten the error message. Not much more needs to be said because
we dump the actual file diff to the screen.