Change how the sgx-types::secs::Secs is instantiated
SECS pages are passed to the enclave creation ioctl() and are never read
from again. Our previous structure assumed that the hardware validated
most of the data in the SECS page. However, it doesn't. It basically
just takes the size parameters and validates those. The rest are filled
in for internal use. The application never reads SECS data once it is
created.

Before this commit, we needed to have the hash of the enclave in order
to create the Secs. This meant that you can't construct and hash the
enclave in parallel. However, now that we know the hardware ignores
`mrenclave` in Secs, we can avoid this dependency. The result is
parallelizable code.