Add the mmap crate
We need a way to do `mmap()`. We need ownership to be tracked properly
and `munmap()` called at the right time. And we need access to the Linux
options. We also need the ability to change the permissions of certain
regions of an existing mapping without taking ownership.

Unfortunately, existing crates either try to be multi-platform
and refuse to support platform-specific options or they are dead (5+
years) and drag in ancient dependencies into the tree.