docs: clarify the attestion steps of ibm/pef
a) payload is generated by encrypting the metadata/hash using a
	randomly-generated symmetric seed, also called as
       	the payload key.
b) a lockbox is generated by encrypting the payload-key using the
	target platforms' public key.
c) the lockbox and the payload and packed togather into a larger
	pack called the ESMoperand.
d) kernel hands over the ESMoperand to ultravisor.
e) ultravisor extracts the lockbox and sends it to the TPM.
f) TPM returns the payload-key to the ultravisor
g) ultraviso extracts the payload, and decrypts it using
	the payload key.
h) ultravisor extracts the hash. Validates it with the
	state of the SVM. On success releases the metadata
	to the kernel.

Signed-off-by: Ram Pai <linuxram@us.ibm.com>