Initial pass at an SGX keep and shim
This doesn't actually accomplish much. However, it does actually
successfully enter and exit an enclave.

The enarx-keep-sgx program takes two arguments: a shim static binary and
a code static binary. These are combined with an SGX TCS page and SSA
pages to form an enclave. The enarx-keep-sgx program then jumps to the
enarx-keep-sgx-shim starting point inside the enclave. Currently, this
does nothing but immediately exit the enclave.