Ensure diagrams are always up-to-date
The Makefile in ./docs/attestation/ is the source of truth containing
the list of diagrams to validate. To this end, a phony helper target was
added to the Makefile to emit the names of the diagrams to check. This
output is meant to be consumed by a script (such as this one).

This helps avoid duplication and any kind of globbing on file extensions
(which means we'd have to remember to update the kinds of file
extensions to glob for, too).