Add the vdso crate
The Linux kernel vDSO is one of the interfaces exposed to userspace by
the kernel. The vDSO is essentially a small ELF library injected into
the process space by the kernel. You can look up functions and call
them.

The driving motivation for this crate is to use the
__vdso_sgx_enter_enclave() function in the vDSO for entering SGX
enclaves.