Implement syscall handlers for exit() and getuid()
This incrementally enables the output and input paths for syscalls. The
exit() syscall handler tests that we can make an outgoing request. It is
also, obviously, necessary for us to terminate the application. The
getuid() syscall handler tests that we can additionally return data to
the syscall call site. Implementation of other syscalls should be fairly
straightforward from here.