Replace enum representation of syscall results
Allowing the microkernel to perform a match expression on a potentially
untrusted page can expose us to undefined behavior if it has been
tampered with.

Replace this with a generic type-constrained structure that will be
written to the page and the microkernel can explicitly read the page
through a pointer of the concrete type it expects.