Introduce minimal messaging interface between keep and shim
In a recent design meeting we decided to try pivoting to a more
minimalist interface between the SEV hypervisor and microkernel to avoid
any accidentally triggering any undefined behavior and also in an effort
to create a low-overhead framework for keep/shim communication.

This is the shape of that framework from our initial meeting.

It is mainly made up of two things:

	1. A `Message` struct, which is the most minimal register
	context we would need to service system calls while still giving
	us the flexibility to define other kinds of service requests.

	2. A `Block` struct, which is a page-sized buffer meant to have
	a `Message` header (the register context) and the rest of its
	space may be used by any data required to carry out that
	request. For example, a proxied `write` syscall would require
	that we store the bytes to be written out in this `Block`, since
	the memory where they actually reside is not accessible to the
	hypervisor.

With this interface, such a request might look a little something like
this:

1. The microkernel traps a system call (let's continue using `write` as
an example).

2. The microkernel recognizes this is a `write` system call and writes
an empty `Block` to the shared syscall page that it uses to communicate
with the hypervisor.

3. It writes the relevant registers that it wants to share to the
`Message` register context struct. In this case:

	rax: 1 (SYS_write)
	rdi: file descriptor to write to
	rsi: a pointer to the bytes to write
	rdx: number of bytes to write

rax, rdi, and rdx are one of the simplests cases, in which simply
writing the values to the registers is sufficient.

rsi is a little more complicated. Simply sharing a pointer into its
protected address space wouldn't work, because the hypervisor cannot
access those pages and therefore that would prohibit its ability to
proxy the syscall.

So, the microkernel will copy the bytes into the `buf` (data) portion of
the `Block` struct and instead of the real address to the bytes, it will
put the address of where the bytes are on the shared syscall page.

	rsi: <address of where the bytes were written to the shared
	page>

Now the `Block` is fully filled out, and contains all of the information
needed to proxy the syscall.

4. The microkernel transfers control to the hypervisor (by executing
HLT, writing to a port, or a VM instruction for yielding)

5. The hypervisor can directly invoke the host-side system call using
the values encoded into the `Block` that resides in the shared syscall
page.

6. After the host finishes with the syscall, the hypervisor can update
the `rax` value on the shared `Block` page to indicate the return value
of the syscall (as expected). Any outgoing parameters or bits of data
will have been modified directly from the proxied syscall as the `Block`
is fully accessible on the hypervisor side.

7. When control shifts back to the microkernel, it may need to propagate
data from the `Block` page as it may have changed during the system
call. For a write call it did not; however, if it was a `read` call then
there are potentially bytes that need to be copied to the protected page
where the application is expecting to see them. Otherwise, it can simply
update `rax` for the process it is servicing and the request has been
proxied.