Fix verify() false positive
OpenSSL's verify() method will only return Err if it experiences some
catastrophic error that prevents it from making a conclusive decision as
to whether the public key's signature can be verified to be on the
digest. It will *NOT* return Err if the key's signature is not on the
digest; it will return Ok(false).

pub fn verify(&self, signature: &[u8]) -> Result<bool, ErrorStack> {
    unsafe {
        let r =
            EVP_DigestVerifyFinal(self.md_ctx, signature.as_ptr() as *mut _, signature.len());
        match r {
            1 => Ok(true),
            0 => {
                ErrorStack::get(); // discard error stack
                Ok(false)
            }
            _ => Err(ErrorStack::get()),
        }
    }
}

We have simply been assuming that if we get Ok from the verify method
that the public key's signature has been found on the digest. We must
inspect the interior `bool` to know whether the signature is verified.

Fixes #520