Simplify Result handling to use std::io::Error
The kvm-ioctl Error type is just a re-export of the Error type exported
by its dependency vmm-sys-util.

The git revision of vmm-sys-util contains a patch that takes advantage
of blanket trait implementations for type conversion, so now the one can
convert seamlessly between the kvm-ioctl Error and std::io::Error.

We can use this to simplify our return type and dispense with boxing a
trait object now.

Closes #556