Fix a crasher
The stack was properly aligned. However, Rust presumed that the stack
was aligned before the call instruction was issued. Because we were
using jmp instead of call, this meant that the output from Rust made
incorrect assumptions about the stack alignment, leading to crashes
later. We solve this by using call rather than jmp so that the Rust
assumptions are correct. This removes the need for a manually aligned
crt0stack which worked around this issue.

There are still some other issues in the crt0stack that need to be
resolved (Harald has fixes for them). However, they don't currently
trigger crashes.

Fixes: #385