Even more conversions for Register
I realized that const conversions to and from Register are safe
precisely because pointer casts are safe. The only thing that is unsafe
is dereferencing a pointer. Therefore, I implemented const pointer and
reference casts.

Likewise, I added a utility for converting a Register to a slice. This
operation is, however, unsafe and is marked as such. It is merely a
convenience function around slice::from_raw_parts().