Fix a crasher in SGX
When libc does its own internal initialization, it calls mmap(). It assumes
this mmap() will succeed and crashes if it fails. The reason that mmap() was
failing was because libc passed -1 as the file descriptor.  However, our
attempt to use a TryInto conversion would fail because it required truncation.
This caused mmap() to return EINVAL.