Add better failure reporting in integration tests
This commit tweaks the standard `assert_eq`s in `integration-tests`
so that instead of only dumping the raw bytes when asserting equality
between expected and given stdio outputs, it now also prints the
mismatch converted to a UTF8 string for easier debugging. The conversion
is lossy so that invalid UTF8 don't trigger false positives in our
testing harness.