Merge shim crates into enarx-keep
This experiment merges `enarx-keep-sgx-shim` and `enarx-keep-sev-shim` into
`enarx-keep`. Then `enarx-keep` grows a build script which builds and installs
the shim binaries into `OUT_DIR`.  Finally, the shim binaries are embedded
into the `enarx-keep` binary via `include_bytes!()`.

The major advantage of this approach is that we can `cargo install enarx-keep`
and the resulting binary is usable. There is no need to fiddle with paths or
other complications. This is a great user experience. It further means we
could merge in `integration-tests` into `enarx-keep` (as a further pull
request) and `enarx-keep` becomes entirely self-contained. It could
potentially be forked off into an independenty repository. Likewise, `cargo
run` can be made to work in a straightforward way.

There are some downsides. First, we don't get any build output from the build
script unless builds fail. Because the shims have dependencies, this could
mean we're waiting on silence for a while. Second, utilities like `cargo fmt`
or `cargo check` wouldn't directly propagate to the shims. So we'd have to
work around this a bit in the CI workflows. The biggest example of this is
`cargo test`. Although integration tests would run, unit tests within the
shims themselves would not run unless you did a manual `cargo test` in the
right directories. Third, this kind of setup might confused some IDEs.
However, I suspect it is no more confusing than our existing `cargo make`
setup; in fact, it might be potentially less confusing.

Co-authored-by: Harald Hoyer <harald@redhat.com>