Add a socket for preattestation
Some technologies, like SEV, SEV-ES and PEF, do attestation during the creation
of the keep rather than afterwards. These technologies need a way to talk to a
client during the keep build process. The general expectation is that if no
socket is specified and the technology does preattestation, it should bring up
the backend either without attestation or with simulated attestation.