Evolve code! macro into impl_const_id!
Moving this into a util module will help with expressing other ioctl
structs that follow the pattern of having a command identifier encoded
into the structure that is sent into the kernel.

Furthermore, the new macro structure allows the consumer to name the
marker trait, its visibility, as well as the type of constant
expression; rather than imposing these choices upon client code. This
adds a better degree of usability to this mechanism.

No functional changes.