Enable SEV protections on VM launch
Because 'cargo test' is multithreaded by default, the SEV launch tests
are mutually exclusive with the SEV unit tests that are gated behind
the "dangerous_tests" feature flag. Manipulating the SEV platform state
during a launch almost always results in a corrupted launch context.

For example: In the time it takes for Thread A to export the certificate
chain to prepare for an SEV launch, another thread has already
manipulated the platform state with a test like platform_reset or
something similar.