feat(tests): Add read_udp test
This test tries to read 65507 in one go.
It fails, if there was a short read.

While this test does not explicitly tests, if the shim has
proxied this to the host with length 65507, an strace shows
that it has.

A future PR will change this to use real UDP.
Related: https://github.com/enarx/enarx-keepldr/issues/154

To test, if bigger reads can be still processed, increase
the `echo` test size to 2MB.

Signed-off-by: Harald Hoyer <harald@redhat.com>