feat(shim-sev): use different interrupt macros
Use the interrupt macro to define the naked assembler function, calling
out to the code inside the macro body.

Also use the `xchg` trick to handle exceptions with an error code.

Also add an integration test with `cpuid`, testing the
`vmm_communication_exception_handler` on SEV-SNP.

Also add a note about the removal of the `idt.rs` module.

Signed-off-by: Harald Hoyer <harald@profian.com>