Don't enter the enclave with disabled exceptions
This approach should provide adequate defense against the [SmashEx]
attack. Basically, we disable exceptions until we enter the Rust
code. If any exception occurs, we loop forever since we cannot trust
the SSA.

In the future, we will update the mainloop to handle this more
gracefully. This will be required whenever we want to handle signals.

[SmashEx]: https://jasonyu1996.github.io/SmashEx/

Signed-off-by: Nathaniel McCallum <nathaniel@profian.com>