Tell rust-analyzer vscode plugin about our crate layout
`rust-analyzer` is pretty awesome, but its crate autodetection only
looks one level deep for Cargo.toml files. This is good enough to find
anything that's mentioned in the toplevel Cargo.toml (like the crates in
`integration/`), but the crates in `internal/` aren't listed because
(for the moment) we have to do special custom stuff to build them.

This commit adds a `.vscode/settings.json` that tells rust-analyzer to
examine the toplevel Cargo.toml and the internal crates (shim-sev,
shim-sgx, and wasmldr. And now rust-analyzer works for those crates.
Hooray!

Signed-off-by: Will Woods <will@profian.com>