v2/guest: Add support for `readv` syscall
Add `validate_iovec_slice{,_mut}` methods on `guest::Platform` trait for
validation of input `iovec` arrays, which is necessary, because the
argument is a dynamically-sized slice and even though the crate could
perform `iovcnt+1` validations to validate the passed iovec slice, there
is no way to store `iovcnt` validated `iovec` structures on the stack,
and, hence, at least `2*iovcnt+1` validations would need to be
performed to, first, compute the sum of length of all `iovec` structures
and allocate data in the block and later to fill the buffers with data
received from the host.
Note, that `iovec` memory layout is identical to Rust slice at the
moment of writing, therefore the trait implementations can validate
the two-dimensional slice in one invocation and return those without any
performance cost.

Signed-off-by: Roman Volosatovs <roman@profian.com>