v2/guest/syscall: Align sockaddr to `align_of::<sockaddr_storage>()`
`lib::sockaddr_storage` is a type designed to fit any of the `sockaddr*`
variants with correct alignment, hence allocate a section with
alignment equal to its alignment.
In practice, the size of the argument should be limited to
`size_of::<sockaddr_storage>()`, but we do not need to enforce it.

Signed-off-by: Roman Volosatovs <roman@profian.com>