Show CPU model in `enarx info`.
- Use constants for CPU manufacturer strings ("AuthenticAMD", "GenuineIntel").
- Create a helper function Vendor::id() for getting the CPU manufacturer string.
- Print the CPU identifier when using the KVM backend.
- Don't use .unwrap() or .transmute(..) in Vendor::get(..)
- Share code for retrieving the CPU identifier (manufacturer string or model name) with CpuId::cpu_identifier(..)

CpuId::cpu_identifier(..) will take in a CpuIdResult created from `__cpuid` with leaf (EAX) `0x80000000` (Get Highest Extended Function Implemented)
and the CPU vendor the backend is designed to use or `None` if it may with with multiple.

Then it will get the manufacturer string to easily determine if the CPU is supported.

If CpuIdResult EAX is less than 0x80000004 then the processor brand string is unavailable, which should be rare with real hardware.
If this happens then [unknown model] will be returned along with the manufacturer string.

Then over a series of `__cpuid` calls with EAX values of `0x80000002`, `0x80000003`, and `0x80000004` the processor brand
string (model name) is retrieved from the EAX, EBX, ECX, and EDX registers (16 bytes at a time).

For example:

0x80000002:

EAX = "Inte"
EBX = "l(R)"
ECX = " Cor"
EDX = "e(TM"

0x80000003:

EAX = ") i5"
EBX = "-469"
ECX = "0K C"
EDX = "PU @"

0x80000004:

EAX = " 3.5"
EBX = "0GHz"
ECX = (all zeroes)
EDX = (all zeroes)

Each register will contain a small portion of the string, by transmuting it to a byte array we can copy
each of the 3 the portions of the string at once.

The final C string will be a max of 48 bytes including a null terminator. It is then parsed as UTF8 and
null characters are removed. It is then returned with the manufacturer string.

A crate `static_assertions` was also added to verify the string length which was derived from some constants in the function.

See the following for additional info:

- https://en.wikipedia.org/wiki/CPUID#EAX=80000002h,80000003h,80000004h:_Processor_Brand_String
- https://doc.rust-lang.org/core/arch/x86_64/fn.__cpuid.html
- https://doc.rust-lang.org/core/arch/x86/struct.CpuidResult.html