shim-sgx: Introduce REPORT microarchitectural data structure
Define a SGX REPORT representation as a Rust data structure but use Intel
SDM naming conventions and explicitly do not enforce documentation. It
makes the whole thing more convoluted, e.g. when comparing your definition
to the Intel SDM. Use only primitive types (integers, arrays) for field
because it is the most portable way to define these structs.

Collect all uarch structs to a single file called sgx_uarch.rs, which
will be located for the moment in the root source directory of shim_sgx,
but will be eventually probably moved to some more "generic" location.

Add a private function encl_ereport() for invoking ENCLU[EREPORT] to the
shim. Mark it with:

This suspends a compiler warning of it yet not having a call site, which
will be addded to get_attestation().

Signed-off-by: Jarkko Sakkinen <jarkko.sakkinen@iki.fi>