feat(enarx): add `--wasmcfgfile` option
The `--wasmcfgfile <FILE>` for the `enarx run` command specifies a
configuration file, which is currently provided as file descriptor 4
for the executable in the keep (wasmldr).

This dynamic keep configuration, is later downloaded by the keep via
network directly to the keep executable after attestation along with the
wasm bytecode.

Signed-off-by: Harald Hoyer <harald@profian.com>