feat(wasmldr): add runtime configuration handling
This patch adds handling of runtime configuration specified in toml
format. Currently the data is read from filedescriptor 4, but in the
future this configuration data is received along with the wasm byte code
after attestation.

In the current configuration file format, the pre-opened files and
sockets are described. Currently only tcp listen sockets and stdio
configuration is handled.

wasmldr inherits stdio filedescriptors as specified and opens tcp listen
sockets as specified. It sets the environment variables according the
the systemd `LISTEN_FDS` specification, so the wasm payload can parse
those.

Signed-off-by: Harald Hoyer <harald@profian.com>