ci: add test for wasm runtime configuration
This test crafts a toml configuration file, which specifies the stdio
inheritance and one tcp listen socket.

It then starts the enarx keep and tries to connect to that listen socket
and succeeds, if it receives the expected data.

This demonstrates functionality of the dynamic runtime configuration, as
well as the new `sock_accept` WASI network functionality.

Signed-off-by: Harald Hoyer <harald@profian.com>