build: read prebuilt internal bins from `ENARX_PREBUILT_*`
The Enarx build script also builds the internal binaries (`shim-sev`,
`shim-sgx`, `wasmldr`) to embed them in the final `enarx` binary.

This approach makes it difficult to build the project if the build phase
does not have any network access. The Nix package manager is a notable
example. Therefore, it is useful to have the option to build the
internal dependencies independently before building `enarx`.

To make that possible, this commit reads the `ENARX_PREBUILT_*`
environment variable when building an internal binary. If the
environment variable is set, it should point to the corresponding
prebuilt binary. In this case, it copies the binary to the expected
output path (`out_bin`).

If the variable is not set, cargo builds the respective binary as
before.

Signed-off-by: Vincent Haupert <mail@vincent-haupert.de>