chore(sev)!: clean up `enarx sev vcek` command
Note well that this is a breaking change. In the previous version of the
command, the VCEK would be mixed with the rest of the Milan certificate
chain and encoded as PEM files. However, none of this is actually
necessary. The platform just needs the DER formatted VCEK somewhere.
From there it is sent to the steward unmodified. The steward already
embeds the rest of the certificate chain.

Signed-off-by: Nathaniel McCallum <nathaniel@profian.com>