fix(sgx): `gdb` heap permission checks
The existing code existing GDB code assumes statically allocated heap with
RWX permissions during the build the time, which does not apply to EDMM
(Enclave Dynamic Memory Management) where only subset of heap is mapped
with variable permissions.

Fix the code by querying from the heap whether the given range is readable
or writable depending on the operations to be applied.

Closes: #1699
Signed-off-by: Jarkko Sakkinen <jarkko@profian.com>