feat: add `enarx deploy` subcommand
`enarx deploy` allows execution of Enarx "packages" from arbitrary URLs.

A package is a tree-like structure, e.g. a directory, containing
an `Enarx.toml` and `main.wasm`.
`enarx deploy` can deploy either a package, using the contained
`Enarx.toml` within its root or a WASM binary directly, in which case
a default config is implicitly used.

```
enarx deploy file:///path/to/my.wasm
```

is a valid invocation, as is

```
enarx deploy file:///path/to/my/package
```

Where `/path/to/my/package` is an absolute path to a local directory
containing (at least) `Enarx.toml` and `main.wasm`

Similarly, `http` and `https` schemes are supported, e.g.

```
enarx deploy https://example.com/path/to/my.wasm
```

If the response of a GET request to specified URL has a `Content-Type` of:
- `application/wasm` => it attempts to execute it
- `application/vnd.drawbridge.directory.v1+json` =>
    it assumes it's a https://github.com/profianinc/drawbridge tree directory
    and attempts to fetch `main.wasm` and `Enarx.toml` from the node rooted at specified URL.

Other media types, and schemes may be supported in the future.

Additionally, this commit adds a two-way communication channel between
the exec layer and the host binary via unnamed Unix sockets. Potentially
this could be used to send e.g. logs or errors to the host if desired. A
TOML-encoded config is sent once on the socket by the host.

Signed-off-by: Roman Volosatovs <roman@profian.com>