fix: specify shim-kvm linker in Cargo.toml rather than Cargo config
By specifying the rustflags in Cargo.toml
we can avoid having to use the Cargo config,
which is preferable as it means that someone using
(for example) `cargo install` won't have to preface the command
with something grotesque like
`CARGO_TARGET_X86_64_UNKNOWN_NONE_RUSTFLAGS="-C linker=gcc"`.

This does mean using a new unstable Cargo feature, but that's
acceptable for the moment as bindeps won't be stabilizing any
time soon, and in the long term we hope to not have to manually specify
the linker for the shims anyway.

Signed-off-by: bstrie <865233+bstrie@users.noreply.github.com>