feat(sgx): `mmap()` handler for the host
Introduce `MmapHost` to provide a host-side `mmap()` handler, which
notifies the host to prepare the guest to do `mmap()`.

In the SGX backend, call `mmap()`.  `PROT_READ | PROT_WRITE` is passed to
`mmap_host()` because **ENCLU[EACCEPTCOPY]** requires read and write
for the virtual memory area.

Finally, adjust the permissions with the new `mprotect_unlocked()` helper,
which is called by the `mmap()` and `mprotect()` guest handlers to perform
the operation, after acquiring the heap.

Signed-off-by: Jarkko Sakkinen <jarkko@profian.com>