build: use new Cargo sparse index support
The index is the list of crate versions from crates.io, which Cargo
uses to resolve dependency versions.
As originally implemented, this index is stored as a Git repo, which
is cloned on first use and pulled for updates.
However, as crates.io has grown the size of the index has become
a bottleneck, and Cargo is preparing to address this by supporting
"sparse indexes", which involves only pulling dependency information
for crates that are actually being used for the compilation.
This promises to be somewhat faster for regular development and
much faster for CI builds.

Signed-off-by: bstrie <865233+bstrie@users.noreply.github.com>