feat: extend the enarx CLI with a key command and change the sign command
Extend the enarx CLI with a `key` command and change the `sign` command.

`-o, --out FILE` is optional and will write the output to a file, otherwise standard output is used.

```console
$ enarx key sgx create --out sgx.key
```

```console
$ enarx key sev create --out sev-author.key
$ enarx key sev create --out sev-id.key
$ enarx key sev sign --sev-author-key sev-author.key --sev-id-key sev-id.key --out sev-id-key-signature.blob
```

`sev-id-key-signature.blob` contains the signature `id_auth.id_key_sig` and the public key of the sev-author.key `id_auth.author_key`,
so the `sev-author.key` does not have to be used in the final signing of the enarx executable.

```console
$ enarx sign --sgx-key sgx.key --sev-id-key sev-id.key --sev-id-key-signature sev-id-key-signature.blob --out sig.json
```

```console
$ enarx key sev digest --key sev-author.key
$ enarx key sev digest --key sev-id.key
$ enarx key sgx digest --key sgx.key
```

Co-authored-by: Harald Hoyer <harald@profian.com>
Signed-off-by: Sriram Tolety <sriram@profian.com>
Signed-off-by: Harald Hoyer <harald@profian.com>