fix: resolve `localhost` to `127.0.0.1`
Typically `localhost` binding is specified in `/etc/hosts`, which
`exec-wasmtime` does not and *should not* have access to.

Eventually, we should remove access to `/etc/resolv.conf` and implement
and use a keep-native DNS resolver. As part of that we could also
resolve `localhost`. Go with a simple solution for now.

Signed-off-by: Roman Volosatovs <roman@profian.com>