fix(cli): handle cannot-be-a-base URLs in `enarx deploy`
`enarx deploy` allows the user to pass either a slug or a URL, and
differentiates between the two based on whether the input successfully
parses as a URL.
However, slugs allow custom hosts to be optionally specified, along
with specific ports, in which the slug looks like `host:1234/user/repo:tag`.
The problem is that this successfully parses as a URL, specifically a
special kind of URL known as a "cannot be a base" URL.
Currently, if a slug like this is used, then enarx deploy will later
panic because what was intended to be the host will instead be
interpreted as a URL scheme.
Instead, this patch causes the input to be interpreted as a slug if
it is parsed as a "cannot be a base" URL.

Signed-off-by: bstrie <865233+bstrie@users.noreply.github.com>