feat(snp): handle PSC requests via GHCB
The kernel will also no longer automatically handle page state change
requests sent as GHCB requests. We need to catch the vmgexit, inspect
the GHCB and execute the page state change requests ourselves.

This requires us to have the GHCB struct in the backend. I can't think
of a good way to shared the definitions between the guest and host
parts, so for now I've just copied over the definitions.

Signed-off-by: Tom Dohrmann <erbse.13@gmx.de>