feat(snp): free up memory for the inactive UPM mapping
When we transition some memory to shared, we don't need the private
mapping anymore, so we can punch a hole into the restricted fd to free
a chunk. Of course that also means that we need to reallocate the
memory when transitioning back to private.

When transitioning memory to private, we can use madvise with
`MADV_FREE` to tell the kernel that the shared mapping is no longer
needed. When memory pressure occurs, the kernel can zero the pages and
free the backing physical memory. Note that this never frees the
virtual memory, so we don't need to reallocate the memory.

Signed-off-by: Tom Dohrmann <erbse.13@gmx.de>