fix: use `from_bits_truncate()`
If enarx or the x86_64 crate does not keep up with CPU tech, using
`from_bits` will prevent the usage of enarx on newer CPUs, because the
checker will bailout here, although it's mostly used as an informative
output.

Happens here with newer XEONs, which have the XTILECFG and XTILEDATA
flags (bit 17 and 18) in Xfrm.

Signed-off-by: Harald Hoyer <harald@hoyer.xyz>