feat(snp): handle hypercall instead of vmgexit
Previously when the keep issued a page state request, the kernel
forwarded that request to userspace which has to decode and handle
that request. With the latest SEV-SNP host patches (v15), the kernel
decodes the PSCs and only asks userspace to handle it by pretending the
keep issued a hypercall.

Signed-off-by: Tom Dohrmann <erbse.13@gmx.de>